+++ slidingbutton/src/main/java/id/ss564/lib/slidingbutton/SlidingButton.kt
@@ -2,28 +2,22 @@ package id.ss564.lib.slidingbutton
 
 import android.animation.Animator
 import android.animation.ValueAnimator
-import android.annotation.SuppressLint
 import android.content.Context
 import android.content.res.ColorStateList
-import android.graphics.PorterDuff
-import android.graphics.PorterDuffColorFilter
-import android.graphics.Typeface
+import android.graphics.*
 import android.graphics.drawable.ColorDrawable
 import android.graphics.drawable.Drawable
 import android.os.Build
 import android.util.AttributeSet
 import android.util.TypedValue
-import android.view.LayoutInflater
-import android.view.MotionEvent
-import android.view.View
+import android.view.*
 import android.view.animation.*
-import android.widget.*
-import androidx.annotation.ColorInt
-import androidx.annotation.Dimension
-import androidx.annotation.DrawableRes
-import androidx.annotation.StringRes
+import android.widget.FrameLayout
+import android.widget.ImageView
+import android.widget.TextView
+import androidx.annotation.*
 import androidx.core.content.ContextCompat
-import androidx.interpolator.view.animation.FastOutSlowInInterpolator
+import kotlin.math.max
 
 /**
  * Created by Anwar on 29 Mar 2020.
@@ -34,16 +28,23 @@ class SlidingButton : FrameLayout {
     private val inflatedView: View
     private lateinit var slidingImage: ImageView
     private lateinit var slidingText: TextView
+    private lateinit var slidingIndicator: View
 
-    private var statusListener: OnStatusChangeListener? = null
+    private var mStateListener: OnStateChangeListener? = null
+    private var slidingListener: OnSlidingListener? = null
+
     private var statusActive = false
         set(value) {
             field = value
-            statusListener?.onStatusChange(value)
+            if (value && showIndicator) animatedIndicator()
+            mStateListener?.onChange(value)
         }
 
     private var startOfButton = 0F
     private var endOfButton = 0F
+
+    private var showIndicator = true
+    private var enableTextAlpha = true
 
     private val stateListIconTint: ColorStateList
 
@@ -89,7 +90,12 @@ class SlidingButton : FrameLayout {
                 slidingText.setTextSize(TypedValue.COMPLEX_UNIT_PX, value)
         }
 
-    val textSize = mTextSize
+    val textSize: Float
+        get() = if (::slidingText.isInitialized) {
+            slidingText.textSize
+        } else {
+            mTextSize
+        }
 
     /**
      * [textPaddings]
@@ -104,7 +110,8 @@ class SlidingButton : FrameLayout {
                 slidingText.setTextColor(value)
         }
 
-    val currentTextColor = if (::slidingText.isInitialized) slidingText.currentTextColor else null
+    val currentTextColor =
+        if (::slidingText.isInitialized) slidingText.currentTextColor else textColors?.defaultColor
 
     var textTypeface: Typeface? = null
         set(value) {
@@ -119,22 +126,17 @@ class SlidingButton : FrameLayout {
                 slidingText.text = value
         }
 
-    /**
-     * [imageSize] is an array have 2 value, index  0 as Width and index 1 as Height
-     */
-    private val imageSize = intArrayOf(0, 0)
-    var buttonWidth
+    var buttonWidth: Int = 0
         set(value) {
-            imageSize[0] = value
+            field = value
             if (::slidingImage.isInitialized) slidingImage.layoutParams.width = value
         }
-        get() = imageSize[0]
-    var buttonHeight
+
+    var buttonHeight: Int = 0
         set(value) {
-            imageSize[1] = value
+            field = value
             if (::slidingImage.isInitialized) slidingImage.layoutParams.height = value
         }
-        get() = imageSize[1]
 
     /**
      * [buttonMargins],[buttonPaddings]
@@ -142,6 +144,45 @@ class SlidingButton : FrameLayout {
      */
     val buttonMargins = intArrayOf(0, 0, 0, 0)
     val buttonPaddings = intArrayOf(0, 0, 0, 0)
+
+    private var trackExtendedTo = TrackExtended.BUTTON
+
+    var cornerRadius: Float = 0F
+        @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
+        set(value) {
+            field = value
+            if (::slidingImage.isInitialized) {
+                RoundedOutlineProvider(value).also {
+                    this.outlineProvider = it
+                    slidingImage.outlineProvider = it
+                    slidingText.outlineProvider = it
+                    slidingIndicator.outlineProvider = it
+                }
+            }
+        }
+
+    var trackBackground: Drawable? = null
+        set(value) {
+            field = value
+            if (::slidingIndicator.isInitialized) {
+                slidingIndicator.background = value
+            }
+        }
+
+    var trackBackgroundTint: ColorStateList? = null
+        set(value) {
+            field = value
+            if (::slidingIndicator.isInitialized) {
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                    slidingIndicator.background?.setTintList(value)
+                } else if (value != null) {
+                    slidingIndicator.background?.colorFilter = PorterDuffColorFilter(
+                        value.getColorForState(slidingIndicator.drawableState, value.defaultColor),
+                        PorterDuff.Mode.SRC_IN
+                    )
+                }
+            }
+        }
 
     constructor(context: Context) : this(context, null)
 
@@ -160,12 +201,21 @@ class SlidingButton : FrameLayout {
         attrs,
         defStyleInt
     ) {
+
         val colorPrimary = TypedValue()
         context.theme.resolveAttribute(R.attr.colorPrimary, colorPrimary, true)
 
+        val ex by lazy {
+            TypedValue().apply {
+                context.theme.resolveAttribute(
+                    R.attr.colorAccent,
+                    this,
+                    true
+                )
+            }.data
+        }
         val colorAccent = TypedValue()
         context.theme.resolveAttribute(R.attr.colorAccent, colorAccent, true)
-
         val arr = context.obtainStyledAttributes(
             attrs,
             R.styleable.SlidingButton,
@@ -180,7 +230,7 @@ class SlidingButton : FrameLayout {
         mTextSize = arr.getDimension(R.styleable.SlidingButton_sliding_text_size, defaultTextSize)
 
         textColors = arr.getColorStateList(R.styleable.SlidingButton_sliding_text_color)
-            ?: ColorStateList.valueOf(colorAccent.data)
+            ?: ColorStateList.valueOf(ex)
 
         mText = arr.getString(R.styleable.SlidingButton_sliding_text)
         textBackground = arr.getDrawable(R.styleable.SlidingButton_sliding_text_background)
@@ -216,24 +266,18 @@ class SlidingButton : FrameLayout {
         buttonIcon = arr.getDrawable(R.styleable.SlidingButton_sliding_button_icon)
             ?: defaultButtonDrawable
 
-        stateListIconTint =
-            arr.getColorStateList(R.styleable.SlidingButton_sliding_button_icon_tint)
-                ?: ColorStateList.valueOf(colorAccent.data)
-        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
-            buttonIcon?.setTintList(stateListIconTint)
-        } else {
-            buttonIcon?.colorFilter =
-                PorterDuffColorFilter(stateListIconTint.defaultColor, PorterDuff.Mode.SRC_IN)
-        }
+        stateListIconTint = arr.getColorStateList(
+            R.styleable.SlidingButton_sliding_button_icon_tint
+        ) ?: ColorStateList.valueOf(colorAccent.data)
 
         buttonBackground = arr.getDrawable(R.styleable.SlidingButton_sliding_button_background)
 
         val defaultButtonSize = resources.getDimensionPixelSize(R.dimen.default_image_height)
-        imageSize[0] = arr.getDimensionPixelSize(
+        buttonWidth = arr.getDimensionPixelSize(
             R.styleable.SlidingButton_sliding_button_width,
             defaultButtonSize
         )
-        imageSize[1] = arr.getDimensionPixelSize(
+        buttonHeight = arr.getDimensionPixelSize(
             R.styleable.SlidingButton_sliding_button_height,
             defaultButtonSize
         )
@@ -282,6 +326,40 @@ class SlidingButton : FrameLayout {
             R.styleable.SlidingButton_sliding_button_paddingBottom,
             0
         )
+
+        enableTextAlpha = arr.getBoolean(R.styleable.SlidingButton_sliding_enabledTextAlpha, true)
+        showIndicator = arr.getBoolean(R.styleable.SlidingButton_sliding_showTrack, false)
+
+        trackBackground = arr.getDrawable(R.styleable.SlidingButton_sliding_trackBackground)
+            ?: ContextCompat.getDrawable(context, R.drawable.default_sliding_indicator_background)
+
+        trackBackgroundTint = arr.getColorStateList(
+            R.styleable.SlidingButton_sliding_trackBackgroundTint
+        )
+
+        val index = arr.getInteger(R.styleable.SlidingButton_sliding_trackExtendTo, 1)
+        trackExtendedTo = arrayOf(TrackExtended.CONTAINER, TrackExtended.BUTTON)[index]
+
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            cornerRadius = arr.getDimension(R.styleable.SlidingButton_sliding_corner_radius, 0F)
+            buttonIcon?.setTintList(stateListIconTint)
+            trackBackground?.setTintList(trackBackgroundTint)
+        } else {
+            buttonIcon?.colorFilter = PorterDuffColorFilter(
+                stateListIconTint.defaultColor,
+                PorterDuff.Mode.SRC_IN
+            )
+            if (trackBackgroundTint != null && trackBackground != null) {
+                trackBackground?.colorFilter = PorterDuffColorFilter(
+                    trackBackgroundTint!!.getColorForState(
+                        trackBackground!!.state,
+                        trackBackgroundTint!!.defaultColor
+                    ),
+                    PorterDuff.Mode.SRC_IN
+                )
+            }
+        }
+
         arr.recycle()
 
         inflatedView = LayoutInflater.from(context).inflate(R.layout.layout_button, this, true)
@@ -290,13 +368,62 @@ class SlidingButton : FrameLayout {
     override fun onFinishInflate() {
         super.onFinishInflate()
         slidingText = inflatedView.findViewById(R.id.slidingText)
+        slidingIndicator = inflatedView.findViewById(R.id.slidingIndicator)
         slidingImage = inflatedView.findViewById(R.id.slidingImage)
 
-        //configure ImageView
+        configureTextView()
+        configureTrackView()
+        configureImageView()
+
+        //Apply Rounded corner to views
+        if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+            RoundedOutlineProvider(cornerRadius).also {
+                this.outlineProvider = it
+                slidingText.outlineProvider = it
+                slidingImage.outlineProvider = it
+                slidingIndicator.outlineProvider = it
+            }
+        }
+
+        configureTouch()
+    }
+
+    private fun configureTrackView() {
+        if (showIndicator) {
+            slidingIndicator.background = trackBackground
+            slidingIndicator.viewTreeObserver.addOnGlobalLayoutListener(object :
+                ViewTreeObserver.OnGlobalLayoutListener {
+                override fun onGlobalLayout() {
+                    slidingIndicator.visibility = View.VISIBLE
+                    val lp = slidingIndicator.layoutParams as LayoutParams
+                    when (trackExtendedTo) {
+                        TrackExtended.BUTTON -> {
+                            lp.width = startOfButton.toInt() + buttonWidth - buttonMargins[2]
+                            lp.height = buttonHeight
+                            lp.marginStart = buttonMargins[0]
+                            lp.marginEnd = buttonMargins[2]
+                        }
+                        TrackExtended.CONTAINER -> {
+                            lp.width = 0
+                            lp.height = this@SlidingButton.measuredHeight
+                            lp.marginStart = 0
+                            lp.marginEnd = 0
+                        }
+                    }
+                    slidingIndicator.layoutParams = lp
+                    inflatedView.viewTreeObserver.removeOnGlobalLayoutListener(this)
+                }
+            })
+        } else {
+            slidingIndicator.visibility = View.GONE
+        }
+    }
+
+    private fun configureImageView() {
         slidingImage.background = buttonBackground
         slidingImage.layoutParams.let { it as LayoutParams }.also {
-            it.width = imageSize[0]
-            it.height = imageSize[1]
+            it.width = buttonWidth
+            it.height = buttonHeight
             it.marginStart = buttonMargins[0]
             it.topMargin = buttonMargins[1]
             it.marginEnd = buttonMargins[2]
@@ -311,8 +438,9 @@ class SlidingButton : FrameLayout {
         )
         slidingImage.scaleType = iconScaleType
         slidingImage.setImageDrawable(buttonIcon)
+    }
 
-        //configure TextView
+    private fun configureTextView() {
         slidingText.background = textBackground
         slidingText.setPaddingRelative(
             textPaddings[0],
@@ -324,22 +452,13 @@ class SlidingButton : FrameLayout {
         slidingText.setTextSize(TypedValue.COMPLEX_UNIT_PX, mTextSize)
         slidingText.setTextColor(textColors)
         slidingText.typeface = textTypeface
-
-        setupSlideTouch()
-        /*if (context.resources.configuration.layoutDirection == View.LAYOUT_DIRECTION_RTL) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-                slidingImage.drawable.isAutoMirrored = true
-            } else {
-                slidingImage.rotationY = 0F
-            }
-        }*/
     }
 
     override fun onSizeChanged(w: Int, h: Int, oldw: Int, oldh: Int) {
         super.onSizeChanged(w, h, oldw, oldh)
         startOfButton = buttonMargins[0].toFloat()
         endOfButton =
-            w.toFloat() - (imageSize[0].toFloat() + buttonMargins[2].toFloat() + paddingEnd.toFloat() + paddingStart.toFloat())
+            w.toFloat() - (buttonWidth.toFloat() + buttonMargins[2].toFloat() + paddingEnd.toFloat() + paddingStart.toFloat())
     }
 
     override fun removeAllViews() = throw IllegalStateException("This method isn't allowed ")
@@ -349,44 +468,43 @@ class SlidingButton : FrameLayout {
     override fun removeViewAt(index: Int) =
         throw IllegalStateException("This method isn't allowed ")
 
-    @SuppressLint("ClickableViewAccessibility")
-    private fun setupSlideTouch() {
-        setOnTouchListener { _, event ->
-            val startTouch = paddingStart + buttonMargins[0]
-            val maxStartTouch = paddingStart + paddingEnd + buttonMargins[0] + imageSize[0]
-            val isStartTouch = event.x in startTouch.toFloat()..maxStartTouch.toFloat()
+    private fun configureTouch() {
+        this.setOnTouchListener { _, event ->
+            val startTouch = (paddingStart + buttonMargins[0]).toFloat()
+            val maxStartTouch = startTouch + buttonWidth
+            val isStartTouch = event.x in startTouch..maxStartTouch
 
-            val maxEndTouch = this.width - paddingEnd - buttonMargins[2]
-            val isEndTouch = event.x in endOfButton..maxEndTouch.toFloat()
+            val maxEndTouch = (this.width - paddingEnd - buttonMargins[2]).toFloat()
+            val isEndTouch = event.x in endOfButton..maxEndTouch
+
             when (event.action) {
                 MotionEvent.ACTION_DOWN -> (isStartTouch && !statusActive) || (isEndTouch && statusActive)
-                MotionEvent.ACTION_MOVE -> {
-                    onMove(event)
-                    true
-                }
-                MotionEvent.ACTION_UP -> {
-                    onUp()
-                    true
-                }
-                else -> false
+                MotionEvent.ACTION_MOVE -> onMove(event, startTouch, endOfButton + buttonWidth)
+                MotionEvent.ACTION_UP -> onUp()
+                else -> true
             }
         }
     }
 
-    private fun onUp() {
-//        val isRtl = context.resources.configuration.layoutDirection == View.LAYOUT_DIRECTION_RTL
-        when {
-            slidingImage.x + slidingImage.width >= this.width * 0.58F -> {
-                animatedToEnd()
-            }
-            slidingImage.x <= startOfButton -> {
-                translateAnimation()
-            }
-            else -> animatedToStart()
+    override fun performClick(): Boolean = super.performClick()
+
+    private fun onUp(): Boolean = when {
+        slidingImage.x + buttonWidth < this.width * 0.55F && slidingImage.x > startOfButton -> {
+            animatedToStart()
+            true
         }
+        slidingImage.x + buttonWidth >= this.width * 0.55F -> {
+            animatedToEnd()
+            true
+        }
+        slidingImage.x <= startOfButton -> {
+            translateAnimation()
+            true
+        }
+        else -> false
     }
 
-    fun changeStatus(active: Boolean, animated: Boolean = false) {
+    fun changeState(active: Boolean, animated: Boolean = false) {
         if (animated && active) {
             statusActive = true
             animatedToEnd()
@@ -407,87 +525,101 @@ class SlidingButton : FrameLayout {
 
         val floatAnimator = ValueAnimator.ofFloat(slidingImage.x, startOfButton)
         floatAnimator.addUpdateListener {
-            slidingImage.x = it.animatedValue as Float
+            updateSlidingXPosition(it.animatedValue as Float)
         }
         floatAnimator.addListener(object : Animator.AnimatorListener {
             override fun onAnimationRepeat(animation: Animator?) {}
 
             override fun onAnimationEnd(animation: Animator?) {
-                if (context.resources.configuration.layoutDirection == View.LAYOUT_DIRECTION_RTL) {
-                    isActivated = true
-                    if (!statusActive) statusActive = true
-                    return
-                }
                 isActivated = false
                 if (statusActive) statusActive = false
             }
 
             override fun onAnimationCancel(animation: Animator?) {}
 
-            override fun onAnimationStart(animation: Animator?) {
-                if (context.resources.configuration.layoutDirection == View.LAYOUT_DIRECTION_RTL) {
-                    isActivated = false
-                }
-            }
+            override fun onAnimationStart(animation: Animator?) {}
         })
         floatAnimator.duration = 115L
-        floatAnimator.interpolator = FastOutSlowInInterpolator()
+        floatAnimator.interpolator = AccelerateDecelerateInterpolator()
         floatAnimator.start()
     }
 
     private fun animatedToEnd() {
         val floatAnimator = ValueAnimator.ofFloat(slidingImage.x, endOfButton)
         floatAnimator.addUpdateListener {
-            slidingImage.x = it.animatedValue as Float
+            updateSlidingXPosition(it.animatedValue as Float)
         }
         floatAnimator.addListener(object : Animator.AnimatorListener {
             override fun onAnimationRepeat(animation: Animator?) {}
 
             override fun onAnimationEnd(animation: Animator?) {
-                if (context.resources.configuration.layoutDirection == View.LAYOUT_DIRECTION_RTL) {
-                    isActivated = false
-                    if (statusActive) statusActive = false
-                    return
-                }
                 isActivated = true
                 if (!statusActive) statusActive = true
             }
 
             override fun onAnimationCancel(animation: Animator?) {}
 
-            override fun onAnimationStart(animation: Animator?) {
-                if (context.resources.configuration.layoutDirection != View.LAYOUT_DIRECTION_RTL) {
-                    isActivated = false
-                }
-            }
+            override fun onAnimationStart(animation: Animator?) {}
         })
         floatAnimator.duration = 115L
-        floatAnimator.interpolator = FastOutSlowInInterpolator()
+        floatAnimator.interpolator = AccelerateDecelerateInterpolator()
         floatAnimator.start()
     }
 
-    private fun onMove(event: MotionEvent) {
+    private fun animatedIndicator() {
+        val animation = ScaleAnimation(
+            0F,
+            1F,
+            1F,
+            1F,
+            slidingIndicator.width.toFloat(),
+            slidingIndicator.height * 0.5F
+        )
+        animation.duration = 500L
+        animation.interpolator = DecelerateInterpolator()
+        slidingIndicator.startAnimation(animation)
+    }
+
+    private fun onMove(event: MotionEvent, start: Float, end: Float): Boolean {
         if (isActivated) isActivated = false
 
-        if (event.x > slidingImage.width / 2
-            && event.x + slidingImage.width / 2 < this.width
-            && (event.x < slidingImage.x + slidingImage.width || slidingImage.x != 0F)
-        ) {
-            slidingImage.x = event.x - slidingImage.width / 2
+        if (event.x < startOfButton + buttonWidth) {
+            updateSlidingXPosition(startOfButton)
+            return true
         }
 
-        if (slidingImage.x + slidingImage.width > this.width
-            && slidingImage.x + slidingImage.width / 2 < this.width
-        ) {
-            slidingImage.x = this.width.toFloat() - slidingImage.width.toFloat()
+        if (event.x in start..end) {
+            updateSlidingXPosition(event.x - buttonWidth)
+            return true
         }
 
-        if (event.x < slidingImage.width / 2 && slidingImage.x > 0) {
-            slidingImage.x = startOfButton
+        return false
+    }
+
+    private fun updateSlidingXPosition(x: Float) {
+        slidingImage.x = x
+        val realX = x - startOfButton
+        val percent = realX / (endOfButton - startOfButton)
+
+        if (enableTextAlpha) {
+            slidingText.alpha = if (percent < 0.2F) 1F - percent else max(0F, 1F - (percent + 0.3F))
         }
+
+        if (showIndicator) {
+            val lp = slidingIndicator.layoutParams as LayoutParams
+            lp.width = when {
+                trackExtendedTo == TrackExtended.CONTAINER && percent == 0F -> 0
+                trackExtendedTo == TrackExtended.CONTAINER && percent < 1F -> x.toInt() + buttonWidth
+                trackExtendedTo == TrackExtended.CONTAINER && percent >= 1F -> x.toInt() + buttonWidth + buttonMargins[2]
+                else -> x.toInt() + buttonWidth - buttonMargins[2]
+            }
+            slidingIndicator.layoutParams = lp
+        }
+        slidingListener?.onSliding(percent)
     }
 
     private fun translateAnimation() {
+        slidingImage.clearAnimation()
         val animation = TranslateAnimation(0F, endOfButton, 0F, 0F)
         animation.interpolator = AccelerateDecelerateInterpolator()
         animation.duration = 350L
@@ -495,6 +627,7 @@ class SlidingButton : FrameLayout {
             override fun onAnimationRepeat(animation: Animation?) {}
 
             override fun onAnimationEnd(animation: Animation?) {
+                slidingImage.clearAnimation()
                 val anim = ScaleAnimation(
                     0.35F,
                     1F,
@@ -505,10 +638,24 @@ class SlidingButton : FrameLayout {
                 )
                 anim.interpolator = DecelerateInterpolator()
                 anim.duration = 225L
+                anim.setAnimationListener(object : Animation.AnimationListener {
+                    override fun onAnimationRepeat(animation: Animation?) {}
+
+                    override fun onAnimationEnd(animation: Animation?) {
+                        slidingImage.clearAnimation()
+                        slidingImage.scaleX = 1F
+                        slidingImage.scaleY = 1F
+                        if (showIndicator) slidingIndicator.alpha = 1F
+                    }
+
+                    override fun onAnimationStart(animation: Animation?) {}
+                })
                 slidingImage.startAnimation(anim)
             }
 
-            override fun onAnimationStart(animation: Animation?) {}
+            override fun onAnimationStart(animation: Animation?) {
+                if (showIndicator) slidingIndicator.alpha = 0F
+            }
         })
         slidingImage.startAnimation(animation)
     }
@@ -517,6 +664,7 @@ class SlidingButton : FrameLayout {
         super.setEnabled(enabled)
         slidingText.isEnabled = enabled
         slidingImage.isEnabled = enabled
+        slidingIndicator.isEnabled = enabled
         changeStateDrawablePreLollipop()
     }
 
@@ -596,14 +744,26 @@ class SlidingButton : FrameLayout {
         )
     }
 
-    fun setOnStatusChangeListener(listener: OnStatusChangeListener?) {
-        statusListener = listener
+    fun setOnStateChangeListener(listener: OnStateChangeListener?) {
+        mStateListener = listener
     }
 
-    fun setOnStatusChangeListener(l: (active: Boolean) -> Unit) {
-        this.setOnStatusChangeListener(object : OnStatusChangeListener {
-            override fun onStatusChange(active: Boolean) {
+    fun setOnStateChangeListener(l: (active: Boolean) -> Unit) {
+        this.setOnStateChangeListener(object : OnStateChangeListener {
+            override fun onChange(active: Boolean) {
                 l.invoke(active)
+            }
+        })
+    }
+
+    fun setOnSlidingListener(listener: OnSlidingListener?) {
+        slidingListener = listener
+    }
+
+    fun setOnSlidingListener(l: (progress: Float) -> Unit) {
+        this.setOnSlidingListener(object : OnSlidingListener {
+            override fun onSliding(progress: Float) {
+                l.invoke(progress)
             }
         })
     }
@@ -622,9 +782,35 @@ class SlidingButton : FrameLayout {
                 stateListIconTint.defaultColor
             ), PorterDuff.Mode.SRC_IN
         )
+
+        trackBackgroundTint?.let {
+            slidingIndicator.background?.colorFilter = PorterDuffColorFilter(
+                it.getColorForState(slidingIndicator.drawableState, it.defaultColor),
+                PorterDuff.Mode.SRC_IN
+            )
+        }
     }
 
-    interface OnStatusChangeListener {
-        fun onStatusChange(active: Boolean)
+    interface OnStateChangeListener {
+        fun onChange(active: Boolean)
+    }
+
+    interface OnSlidingListener {
+        fun onSliding(progress: Float)
+    }
+
+    @RequiresApi(Build.VERSION_CODES.LOLLIPOP)
+    private class RoundedOutlineProvider(val radius: Float) : ViewOutlineProvider() {
+        override fun getOutline(view: View, outline: Outline) {
+            val rect = Rect(0, 0, view.width, view.height)
+            outline.setRoundRect(rect, radius)
+            view.clipToOutline = true
+        }
+    }
+
+    @Keep
+    enum class TrackExtended constructor(val value: Int) {
+        CONTAINER(0),
+        BUTTON(1)
     }
 }