+++ CHANGELOG.md
@@ -1,7 +1,74 @@
 # Change Log
 
+
+
+## Version 2.0.0 _(15 Apr 2020)_
+
+Behavior change on state listener, rename to `OnStateChangeListener`. So you can listen *state change* using method `setOnStateChangeListener()` with value `SlidingButton.OnStateChangeListener`.
+
+Change state rename to `slidingButton.changeState(active,animated)`
+
+
+
+Additional listener when button sliding. You can now listen it using method `setOnSlidingListener()` with value `SlidingButton.OnSlidingListener`. For example :
+
+- Java
+
+  ```java
+  slidingButton.setOnSlidingListener(new SlidingButton.OnSlidingListener{
+      
+      @Override
+      public void onChange(float progress){
+          // do what you wanna do
+      }
+  });
+  ```
+
+  
+
+- Kotlin
+
+  ```kotlin
+  slidingButton.setOnSlidingListener { progress ->
+      // do what you wanna do
+  }
+  ```
+
+
+
+Additional features like **text alpha** when sliding the button, **track indicator** and **corner radius** (especially for API 21). You can set using attributes like the following :
+
+- `app:sliding_enabledTextAlpha`
+
+  Use to alpha text when sliding, value of this attribute **Boolean**. Default value true
+
+- `app:sliding_showTrack`
+
+  Use to show track indicator when sliding, value of this attribute **Boolean**. Default value false
+
+- `app:sliding_trackBackground`
+
+  Use to set track indicator background.
+
+- `app:sliding_trackBackgroundTint`
+
+  Use to set tint track indicator background.
+
+- `app:sliding_trackExtendTo`
+
+  Use to set track extended to `container` or `button`
+
+- `app:sliding_corner_radius` only on API level 21 (Lollipop)
+
+
+
+-------
+
 ## Version 1.0.2 _(09 Apr 2020)_
+
 Add method to change status programmatically. So you can change status use method `changeStatus(active,animated)` as you needed
+
+
 
 -------
 
@@ -16,11 +83,14 @@ repositories {
 }
 ```
 
+
+
 -------
 
 ## Version 1.0.0 _(03 Apr 2020)_
 
 Slide button library for android, we hope this library is useful and easy to customize as you needed.
+
 
 
 -------